Add integration test for name collisions
